Managed to create a whole JFrame and pane inside of it. Inside of that pane I created a label that holds an object called textArea which creates the text.
I managed to change the pane background color to Dark Gray, and change the font, type and color of the textArea text.

My next attempt is going to be to create a button 'Play' that allows the user to move on to the next 'screen' and actually start playing the game.

And I will try to position the text to the correct position, preferably the center as it is touching the left screen 'wall' at the moment.

The button 'Play' should set the existing Gui to .setVisible(false) and should set the next labels to .setVisible(true) at least I think that it should do this not sure yet cuz I haven't actually tried it.

Modestas Cepulis T00199237
06/11/2018