Okaaaay, so I did a lot on this class. :D
First of all I've created a new class called MyGuiTest which I used to test my new design of the JFrame and so that I could actually figure it out how to do all of the positioning and stuff.

What helped my with positioning was .setBounds and .setLayout(false).
Once I managed to figure out what was the problem, I moved on to putting a blue background on the labels just to make sure that I know what size they are.

After that I changed all the text font to 'New Times Roman' (probably, - definitely I'm going to be changing that), changed the font size and I set it to be PLAIN.

When I had my text in place I started to work on the buttons. I've made 4 buttons for different actions, which are as stands:
'Start New Game' - Which will allow the user to start a new game, etc - begin playing the actual game,
'Options' - I'm not sure what will this allow the user yet, but it might be something like: 'Change Volume' which would allow the user to change the volume of the music/sounds, 'Change Text Color' - which would allow the user to change the color of the main text, etc...
'About the game' - This part of the section will give the user all of the information on how to play and use the game features.
'Exit' - Will allow the user to exit out of the game.

When I was done with the buttons, I created different methods that create different panels whenever the user click one of the buttons. (For now the panels are all the same, but later on I will change them)

To make everything clear I have declared my objects and variables at the top of the code, created different user-defined methods for creating the panels and the buttons.

For my next class I will try to attempt to make the 'Scenes' for the 3 last buttons. (Exit, About the Game and Options).

As for now I don't know what I'm going to do with the story of the game. I will have to plan that out later on.

T00199237 Modestas Cepulis
08/11/2018