Started to do the first "interactions" with the game. Managed to make different buttons for different choices, as well as the health and the current weapon of the player.
Now the user is able to move betweent few 'text' scenes and come back to the previous one, as for now there is only 2 of them.

I was working on the 'exit the game' screen which will ask the user if he really wants to exit the game and if the user presses 'yes' the game is going to end and if the user is going to press 'no' the screen just sets the 'main menu' panels.

Now for the next aim I will try to make that the user can enter the name of the character, chooses an avatar or something like that.

Of course as I need to finish making 'about the game' and 'options' screens before I can move to the ones that I've mentioned above.

Modestas Cepulis T00199237
12/11/2018