Managed to finish the 'exit' button and currently working on the INVENTORY and MAP which can be accessed by pressing 'I' - for inventory and 'M' - for map.
For map I'm not sure yet if it is going to be accessed from the beginning or the player will need to find it later in the game, but the map would show the 'world' areas.

As for inventory, the inventory should show current health, current level, current weapon and current items that the player has. (Maybe quests as well, not sure yet about that)

Later on I will have to finish the options button, I will try to work on that on the next class.

Modestas Cepulis T00199237
12/11/2018