Finally finished the whole main menu UI, now every buttons works and you can go back to the main menu from every button.
Change the colour scheme so it actually looks decent instead of having blue boxes around it.

The options only changes the containers background color, it changes the text colour but only for options scene (might leave it as it is)

Now finally I can start working on the actual game. I was planning to do it like this: First show some text, etc, you wake up in a cave, then show the caves picture, and then show the options that the user can do, etc - look around, go the the corner. . .

Modestas Cepulis T00199237
15/11/2018