Managed to finish up the story line to that point where players meet someone to fight.
fighting system works in away that it gets a random number between a certain range and puts it in 'playerAttack' or 'enemyAttack' attributes.
And then subtracts the health of the player or enemy depending on what the number is.

for enemy attack weapons Ive made an array which contains different named weapons that the guard can use for attack.

for the next class I will try to make a shopping system that the user could purchase or sell items to the market.

Modestas Cepulis T00199237
21/11/2018